engine: extract esm/{plot_tile,star}.js
