+++ engine/esm/transpiled.js
@@ -135,9 +135,11 @@ import {
 } from "./interfaces.js";
 
 import { Annotation, Circle, Poly, PolyLine } from "./annotation.js";
+import { BasePlanets } from "./baseplanets.js";
 import { CameraParameters } from "./camera_parameters.js";
 import { WebFile } from "./web_file.js";
 import { UiTools } from "./ui_tools.js";
+import { Star, Galaxy } from "./star.js";
 
 import { ColorMapContainer } from "./layers/color_map_container.js";
 import { WcsImage } from "./layers/wcs_image.js";
@@ -149,6 +151,7 @@ import { RenderTriangle } from "./render_triangle.js";
 import { EquirectangularTile } from "./equirectangular_tile.js";
 import { HealpixTile } from "./healpix_tile.js";
 import { MercatorTile } from "./mercator_tile.js";
+import { PlotTile } from "./plot_tile.js";
 
 
 // wwtlib.ScaleTypes
@@ -11280,15 +11283,15 @@ Planets._drawPlanet3d = function (renderContext, planetID, centerPoint) {
       renderContext.lighting = oldLighting;
     } else {
       if (!planetID) {
-        Planets.drawPointPlanet(renderContext, new Vector3d(), (10 * Planets._planetDiameters[planetID]), Planets.planetColors[planetID], true);
+        BasePlanets.drawPointPlanet(renderContext, new Vector3d(), (10 * Planets._planetDiameters[planetID]), Planets.planetColors[planetID], true);
       }
       else if (planetID < 9 || planetID === 19) {
         var size = (800 * Planets._planetDiameters[planetID]);
-        Planets.drawPointPlanet(renderContext, new Vector3d(), Math.max(0.05, Math.min(0.1, size)), Planets.planetColors[planetID], true);
+        BasePlanets.drawPointPlanet(renderContext, new Vector3d(), Math.max(0.05, Math.min(0.1, size)), Planets.planetColors[planetID], true);
       }
       else if (sizeIndexParam > 0.002) {
         var size = (800 * Planets._planetDiameters[planetID]);
-        Planets.drawPointPlanet(renderContext, new Vector3d(), Math.max(0.05, Math.min(0.1, size)), Planets.planetColors[planetID], true);
+        BasePlanets.drawPointPlanet(renderContext, new Vector3d(), Math.max(0.05, Math.min(0.1, size)), Planets.planetColors[planetID], true);
       }
     }
   }
@@ -11425,42 +11428,6 @@ Planets._initRings = function () {
   Planets._ringsVertexBuffer.unlock();
 };
 
-Planets.drawPointPlanet = function (renderContext, location, size, color, zOrder) {
-  var center = location;
-  var rad = size / 2;
-  if (renderContext.gl != null) {
-    var ppList = new PointList(renderContext);
-    ppList.minSize = 20;
-    ppList.addPoint(location.copy(), color._clone(), new Dates(0, 1), size / 100);
-    ppList.depthBuffered = true;
-    ppList.draw(renderContext, 1, false);
-  }
-  else {
-    var screenSpacePnt = renderContext.WVP.transform(center);
-    if (screenSpacePnt.z < 0) {
-      return;
-    }
-    if (!zOrder) {
-      if (Vector3d.dot(renderContext.get_viewPoint(), center) < 0.55) {
-        return;
-      }
-    }
-    var ctx = renderContext.device;
-    ctx.save();
-    ctx.beginPath();
-    ctx.arc(screenSpacePnt.x, screenSpacePnt.y, rad, 0, Math.PI * 2, true);
-    ctx.lineWidth = 1;
-    ctx.fillStyle = color.toString();
-    if (true) {
-      ctx.fill();
-    }
-    ctx.globalAlpha = 1;
-    ctx.strokeStyle = color.toString();
-    ctx.stroke();
-    ctx.restore();
-  }
-};
-
 Planets.getAdjustedPlanetRadius = function (planetID) {
   if (planetID > Planets._planetDiameters.length - 1) {
     planetID = 19;
@@ -11489,7 +11456,7 @@ Planets._drawPlanet = function (renderContext, planetID, opacity) {
   if (((planetID < 14) && Planets._planetScales[planetID] < (renderContext.viewCamera.zoom / 6) / 400)) {
     if (planetID < 10 || ((planetID < 14) && Planets._planetScales[planetID] > (renderContext.viewCamera.zoom / 6) / 6400)) {
       var point = Coordinates.raDecTo3d(planetPosition.RA, planetPosition.dec);
-      Planets.drawPointPlanet(renderContext, point, 3, Planets.planetColors[planetID], false);
+      BasePlanets.drawPointPlanet(renderContext, point, 3, Planets.planetColors[planetID], false);
     }
     return;
   }
@@ -14486,213 +14453,6 @@ SpaceTimeController.dateToJD = function (Year, Month, Day, bGregorianCalendar) {
 var SpaceTimeController$ = {};
 
 registerType("SpaceTimeController", [SpaceTimeController, SpaceTimeController$, null]);
-
-// wwtlib.Star
-
-export function Star(input) {
-  this.magnitude = 0;
-  this.RA = 0;
-  this.dec = 0;
-  this.BMV = 0;
-  this.id = 0;
-  this.absoluteMagnitude = 0;
-  this.par = 0;
-  this.distance = 0;
-  var sa = input.split('\t');
-  this.id = parseInt(ss.replaceString(sa[0], 'HIP', ''));
-  this.dec = parseFloat(sa[3]);
-  this.RA = parseFloat(sa[2]) / 15;
-  if (sa.length > 4) {
-    try {
-      if (sa[4].toUpperCase() !== 'NULL' && !!sa[4]) {
-        this.magnitude = parseFloat(sa[4]);
-      }
-    }
-    catch ($e1) {
-    }
-  }
-  if (sa.length > 5) {
-    try {
-      this.BMV = parseFloat(sa[5]);
-      this._makeColor(this.BMV);
-    }
-    catch ($e2) {
-    }
-  }
-  if (sa.length > 6) {
-    this.par = parseFloat(sa[6]);
-    this._makeDistanceAndMagnitude();
-  }
-}
-
-var Star$ = {
-  get_name: function () {
-    return 'HIP' + this.id.toString();
-  },
-
-  get_coordinates: function () {
-    return Coordinates.fromRaDec(this.RA, this.dec);
-  },
-
-  get_asPlace: function () {
-    var place = Place.create(this.get_name(), this.dec, this.RA, 1, Constellations.containment.findConstellationForPoint(this.RA, this.dec), 4, -1);
-    place.set_magnitude(this.magnitude);
-    place.set_distance(this.distance);
-    return place;
-  },
-
-  stars: function (input, newish) {
-    var sa = input.split('\t');
-    this.id = parseInt(sa[0]);
-    this.RA = parseFloat(sa[1]) / 15;
-    this.dec = parseFloat(sa[2]);
-    if (sa.length > 3) {
-      try {
-        this.magnitude = parseFloat(sa[3]);
-      }
-      catch ($e1) {
-      }
-    }
-    if (sa.length > 4) {
-      try {
-        this.col = Color.load(sa[4]);
-      }
-      catch ($e2) {
-      }
-    }
-  },
-
-  _makeDistanceAndMagnitude: function () {
-    this.distance = 1 / (this.par / 1000);
-    this.absoluteMagnitude = this.magnitude - 5 * (Util.logN(this.distance, 10) - 1);
-    this.distance *= 206264.806;
-  },
-
-  _makeColor: function (bmv) {
-    var c = 4294967295;
-    if (bmv <= -0.32) {
-      c = 4288854271;
-    } else if (bmv <= -0.31) {
-      c = 4288919807;
-    } else if (bmv <= -0.3) {
-      c = 4288985855;
-    }
-    // mistake in original source. Should this be "<= -0.29" or something?
-    // else if (bmv <= -0.3) {
-    //   c = 4289051391;
-    // }
-    else if (bmv <= -0.28) {
-      c = 4289182975;
-    } else if (bmv <= -0.26) {
-      c = 4289314303;
-    } else if (bmv <= -0.24) {
-      c = 4289445887;
-    } else if (bmv <= -0.2) {
-      c = 4289708799;
-    } else if (bmv <= -0.16) {
-      c = 4290037503;
-    } else if (bmv <= -0.14) {
-      c = 4290169087;
-    } else if (bmv <= -0.12) {
-      c = 4290366207;
-    } else if (bmv <= -0.09) {
-      c = 4290563583;
-    } else if (bmv <= -0.06) {
-      c = 4290892031;
-    } else if (bmv <= 0) {
-      c = 4291483391;
-    } else if (bmv <= 0.06) {
-      c = 4292009215;
-    } else if (bmv <= 0.14) {
-      c = 4292732159;
-    } else if (bmv <= 0.19) {
-      c = 4293126399;
-    } else if (bmv <= 0.31) {
-      c = 4294111999;
-    } else if (bmv <= 0.36) {
-      c = 4294571775;
-    } else if (bmv <= 0.43) {
-      c = 4294965756;
-    } else if (bmv <= 0.54) {
-      c = 4294964979;
-    } else if (bmv <= 0.59) {
-      c = 4294964203;
-    } else if (bmv <= 0.63) {
-      c = 4294963687;
-    } else if (bmv <= 0.66) {
-      c = 4294963169;
-    } else if (bmv <= 0.74) {
-      c = 4294962909;
-    } else if (bmv <= 0.82) {
-      c = 4294961877;
-    } else if (bmv <= 0.92) {
-      c = 4294960324;
-    } else if (bmv <= 1.15) {
-      c = 4294959032;
-    } else if (bmv <= 1.3) {
-      c = 4294958516;
-    } else if (bmv <= 1.41) {
-      c = 4294955933;
-    } else if (bmv <= 1.48) {
-      c = 4294954385;
-    } else if (bmv <= 1.52) {
-      c = 4294953351;
-    } else if (bmv <= 1.55) {
-      c = 4294952319;
-    } else if (bmv <= 1.56) {
-      c = 4294951287;
-    } else if (bmv <= 1.61) {
-      c = 4294950257;
-    } else if (bmv <= 1.72) {
-      c = 4294948966;
-    } else if (bmv <= 1.84) {
-      c = 4294947419;
-    } else if (bmv <= 2) {
-      c = 4294946129;
-    }
-    this.col = Color.fromInt(c);
-  }
-};
-
-registerType("Star", [Star, Star$, null]);
-
-// wwtlib.Galaxy
-
-export function Galaxy(br) {
-  this.RA = 0;
-  this.dec = 0;
-  this.distance = 0;
-  this.type = 0;
-  this.eTypeBucket = 0;
-  this.size = 5;
-  this.sdssID = 0;
-  this.sdssID = br.readInt64();
-  this.RA = br.readSingle();
-  this.dec = br.readSingle();
-  this.distance = br.readSingle();
-  this.eTypeBucket = br.readByte();
-  this.size = br.readSingle();
-}
-
-Galaxy._eTypeBuckets = [-3, -0.186, -0.168, -0.158, -0.15, -0.143, -0.137, -0.13, -0.123, -0.115, -0.104, -0.089, -0.068, -0.042, -0.011, 0.024, 0.064, 0.111, 0.169, 0.252, 3];
-
-Galaxy.getEType = function (value) {
-  var a = 0;
-  var b = Galaxy._eTypeBuckets.length - 1;
-  while (b - a > 1) {
-    var m = (a + b) / 2;
-    if (value > Galaxy._eTypeBuckets[m]) {
-      a = m;
-    } else {
-      b = m;
-    }
-  }
-  return a;
-};
-
-var Galaxy$ = {};
-
-registerType("Galaxy", [Galaxy, Galaxy$, null]);
 
 // wwtlib.LatLngEdges
 
@@ -26068,7 +25828,7 @@ var Object3dLayer$ = {
     var rotation = Matrix3d.multiplyMatrix(Matrix3d.multiplyMatrix(Matrix3d._rotationZ(-this._roll$1 / 180 * Math.PI), Matrix3d._rotationX(-this._pitch$1 / 180 * Math.PI)), Matrix3d._rotationY(this._heading$1 / 180 * Math.PI));
     renderContext.set_world(Matrix3d.multiplyMatrix(Matrix3d.multiplyMatrix(Matrix3d.multiplyMatrix(rotation, Matrix3d._scaling(this._scale$1.x, this._scale$1.y, this._scale$1.z)), Matrix3d.translation(this._translate$1)), oldWorld));
     renderContext.set_twoSidedLighting(this.get_twoSidedGeometry());
-    Planets.drawPointPlanet(renderContext, new Vector3d(), 1, Colors.get_red(), false);
+    BasePlanets.drawPointPlanet(renderContext, new Vector3d(), 1, Colors.get_red(), false);
     if (this._lightID$1 > 0) {
     } else {
       if (this.object3d != null) {
@@ -30367,249 +30127,6 @@ var VoTableLayer$ = {
 };
 
 registerType("VoTableLayer", [VoTableLayer, VoTableLayer$, Layer]);
-
-// wwtlib.PlotTile
-
-export function PlotTile() {
-  this._topDown$1 = true;
-  this.backslash = false;
-  this._vertexList$1 = null;
-  this._childTriangleList$1 = null;
-  this._stars$1 = [];
-  this._subDivisionLevel$1 = 4;
-  this._subDivided$1 = false;
-  Tile.call(this);
-}
-
-PlotTile.create = function (level, xc, yc, dataset, parent) {
-  var temp = new PlotTile();
-  temp.parent = parent;
-  temp.level = level;
-  temp.tileX = xc;
-  temp.tileY = yc;
-  temp.dataset = dataset;
-  temp._topDown$1 = !dataset.get_bottomsUp();
-  if (temp.tileX !== xc) {
-    alert('bad');
-  }
-  if (!!dataset.get_meanRadius()) {
-    temp.set__demScaleFactor(dataset.get_meanRadius());
-  }
-  else {
-    if (!dataset.get_dataSetType()) {
-      temp.set__demScaleFactor(6371000);
-    } else {
-      temp.set__demScaleFactor(3396010);
-    }
-  }
-  temp.computeBoundingSphere();
-  return temp;
-};
-
-var PlotTile$ = {
-  computeBoundingSphere: function () {
-    this._initializeGrids$1();
-    this.topLeft = this.bounds[0 + 3 * 0].position.copy();
-    this.bottomRight = this.bounds[2 + 3 * 2].position.copy();
-    this.topRight = this.bounds[2 + 3 * 0].position.copy();
-    this.bottomLeft = this.bounds[0 + 3 * 2].position.copy();
-    this.calcSphere();
-  },
-
-  renderPart: function (renderContext, part, opacity, combine) {
-    if (renderContext.gl != null) {
-    } else {
-      if (!part) {
-        var $enum1 = ss.enumerate(this._stars$1);
-        while ($enum1.moveNext()) {
-          var star = $enum1.current;
-          var radDec = 25 / Math.pow(1.6, star.magnitude);
-          Planets.drawPointPlanet(renderContext, star.position, radDec, star.col, false);
-        }
-      }
-    }
-  },
-
-  requestImage: function () {
-    if (!this.downloading && !this.readyToRender) {
-      this.downloading = true;
-      this._webFile$1 = new WebFile(URLHelpers.singleton.rewrite(this.get_URL(), 0));
-      this._webFile$1.onStateChange = ss.bind('fileStateChange', this);
-      this._webFile$1.send();
-    }
-  },
-
-  fileStateChange: function () {
-    if (this._webFile$1.get_state() === 2) {
-      this.downloading = false;
-      this.readyToRender = false;
-      this.errored = true;
-      this.requestPending = false;
-      tileCacheRemoveFromQueue(this.get_key(), true);
-    } else if (this._webFile$1.get_state() === 1) {
-      this.texReady = true;
-      this.downloading = false;
-      this.errored = false;
-      this.readyToRender = this.texReady && (this.demReady || !this.demTile);
-      this.requestPending = false;
-      tileCacheRemoveFromQueue(this.get_key(), true);
-      this._loadData$1(this._webFile$1.getText());
-    }
-  },
-  _loadData$1: function (data) {
-    var rows = ss.replaceString(data, '\r\n', '\n').split('\n');
-    var firstRow = true;
-    var type = 0;
-    var star = null;
-    var $enum1 = ss.enumerate(rows);
-    while ($enum1.moveNext()) {
-      var row = $enum1.current;
-      if (firstRow) {
-        firstRow = false;
-        continue;
-      }
-      if (ss.trim(row).length > 5) {
-        star = new Star(row);
-        star.position = Coordinates.raDecTo3dAu(star.RA, star.dec, 1);
-        this._stars$1.push(star);
-      }
-    }
-  },
-
-  isPointInTile: function (lat, lng) {
-    if (!this.level) {
-      return true;
-    }
-    if (this.level === 1) {
-      if ((lng >= 0 && lng <= 90) && (!this.tileX && this.tileY === 1)) {
-        return true;
-      }
-      if ((lng > 90 && lng <= 180) && (this.tileX === 1 && this.tileY === 1)) {
-        return true;
-      }
-      if ((lng < 0 && lng >= -90) && (!this.tileX && !this.tileY)) {
-        return true;
-      }
-      if ((lng < -90 && lng >= -180) && (this.tileX === 1 && !this.tileY)) {
-        return true;
-      }
-      return false;
-    }
-    if (!this.demReady || this.demData == null) {
-      return false;
-    }
-    var testPoint = Coordinates.geoTo3dDouble(-lat, lng);
-    var top = this._isLeftOfHalfSpace$1(this.topLeft.copy(), this.topRight.copy(), testPoint);
-    var right = this._isLeftOfHalfSpace$1(this.topRight.copy(), this.bottomRight.copy(), testPoint);
-    var bottom = this._isLeftOfHalfSpace$1(this.bottomRight.copy(), this.bottomLeft.copy(), testPoint);
-    var left = this._isLeftOfHalfSpace$1(this.bottomLeft.copy(), this.topLeft.copy(), testPoint);
-    if (top && right && bottom && left) {
-      return true;
-    }
-    return false;
-  },
-  _isLeftOfHalfSpace$1: function (pntA, pntB, pntTest) {
-    pntA.normalize();
-    pntB.normalize();
-    var cross = Vector3d.cross(pntA, pntB);
-    var dot = Vector3d.dot(cross, pntTest);
-    return dot < 0;
-  },
-  _initializeGrids$1: function () {
-    this._vertexList$1 = [];
-    this._childTriangleList$1 = new Array(4);
-    this._childTriangleList$1[0] = [];
-    this._childTriangleList$1[1] = [];
-    this._childTriangleList$1[2] = [];
-    this._childTriangleList$1[3] = [];
-    this.bounds = new Array(9);
-    if (this.level > 0) {
-      if (this.parent == null) {
-        this.parent = tileCacheGetTile(this.level - 1, this.tileX / 2, this.tileY / 2, this.dataset, null);
-      }
-      var parent = this.parent;
-      var xIndex = this.tileX % 2;
-      var yIndex = this.tileY % 2;
-      if (this.level > 1) {
-        this.backslash = parent.backslash;
-      }
-      else {
-        this.backslash = (xIndex === 1 ^ yIndex === 1) === 1;
-      }
-      this.bounds[0 + 3 * 0] = parent.bounds[xIndex + 3 * yIndex].copy();
-      this.bounds[1 + 3 * 0] = this._midpoint$1(parent.bounds[xIndex + 3 * yIndex], parent.bounds[xIndex + 1 + 3 * yIndex]);
-      this.bounds[2 + 3 * 0] = parent.bounds[xIndex + 1 + 3 * yIndex].copy();
-      this.bounds[0 + 3 * 1] = this._midpoint$1(parent.bounds[xIndex + 3 * yIndex], parent.bounds[xIndex + 3 * (yIndex + 1)]);
-      if (this.backslash) {
-        this.bounds[1 + 3 * 1] = this._midpoint$1(parent.bounds[xIndex + 3 * yIndex], parent.bounds[xIndex + 1 + 3 * (yIndex + 1)]);
-      }
-      else {
-        this.bounds[1 + 3 * 1] = this._midpoint$1(parent.bounds[xIndex + 1 + 3 * yIndex], parent.bounds[xIndex + 3 * (yIndex + 1)]);
-      }
-      this.bounds[2 + 3 * 1] = this._midpoint$1(parent.bounds[xIndex + 1 + 3 * yIndex], parent.bounds[xIndex + 1 + 3 * (yIndex + 1)]);
-      this.bounds[0 + 3 * 2] = parent.bounds[xIndex + 3 * (yIndex + 1)].copy();
-      this.bounds[1 + 3 * 2] = this._midpoint$1(parent.bounds[xIndex + 3 * (yIndex + 1)], parent.bounds[xIndex + 1 + 3 * (yIndex + 1)]);
-      this.bounds[2 + 3 * 2] = parent.bounds[xIndex + 1 + 3 * (yIndex + 1)].copy();
-      this.bounds[0 + 3 * 0].tu = 0 * tileUvMultiple;
-      this.bounds[0 + 3 * 0].tv = 0 * tileUvMultiple;
-      this.bounds[1 + 3 * 0].tu = 0.5 * tileUvMultiple;
-      this.bounds[1 + 3 * 0].tv = 0 * tileUvMultiple;
-      this.bounds[2 + 3 * 0].tu = 1 * tileUvMultiple;
-      this.bounds[2 + 3 * 0].tv = 0 * tileUvMultiple;
-      this.bounds[0 + 3 * 1].tu = 0 * tileUvMultiple;
-      this.bounds[0 + 3 * 1].tv = 0.5 * tileUvMultiple;
-      this.bounds[1 + 3 * 1].tu = 0.5 * tileUvMultiple;
-      this.bounds[1 + 3 * 1].tv = 0.5 * tileUvMultiple;
-      this.bounds[2 + 3 * 1].tu = 1 * tileUvMultiple;
-      this.bounds[2 + 3 * 1].tv = 0.5 * tileUvMultiple;
-      this.bounds[0 + 3 * 2].tu = 0 * tileUvMultiple;
-      this.bounds[0 + 3 * 2].tv = 1 * tileUvMultiple;
-      this.bounds[1 + 3 * 2].tu = 0.5 * tileUvMultiple;
-      this.bounds[1 + 3 * 2].tv = 1 * tileUvMultiple;
-      this.bounds[2 + 3 * 2].tu = 1 * tileUvMultiple;
-      this.bounds[2 + 3 * 2].tv = 1 * tileUvMultiple;
-    } else {
-      this.bounds[0 + 3 * 0] = PositionTexture.create(0, -1, 0, 0, 0);
-      this.bounds[1 + 3 * 0] = PositionTexture.create(0, 0, 1, 0.5, 0);
-      this.bounds[2 + 3 * 0] = PositionTexture.create(0, -1, 0, 1, 0);
-      this.bounds[0 + 3 * 1] = PositionTexture.create(-1, 0, 0, 0, 0.5);
-      this.bounds[1 + 3 * 1] = PositionTexture.create(0, 1, 0, 0.5, 0.5);
-      this.bounds[2 + 3 * 1] = PositionTexture.create(1, 0, 0, 1, 0.5);
-      this.bounds[0 + 3 * 2] = PositionTexture.create(0, -1, 0, 0, 1);
-      this.bounds[1 + 3 * 2] = PositionTexture.create(0, 0, -1, 0.5, 1);
-      this.bounds[2 + 3 * 2] = PositionTexture.create(0, -1, 0, 1, 1);
-    }
-  },
-  _midpoint$1: function (positionNormalTextured, positionNormalTextured_2) {
-    var a1 = Vector3d.lerp(positionNormalTextured.position, positionNormalTextured_2.position, 0.5);
-    var a1uv = Vector2d.lerp(Vector2d.create(positionNormalTextured.tu, positionNormalTextured.tv), Vector2d.create(positionNormalTextured_2.tu, positionNormalTextured_2.tv), 0.5);
-    a1.normalize();
-    return PositionTexture.createPos(a1, a1uv.x, a1uv.y);
-  },
-
-  createGeometry: function (renderContext) {
-    if (this.geometryCreated) {
-      return true;
-    }
-    this.geometryCreated = true;
-    Tile.prototype.createGeometry.call(this, renderContext);
-    return true;
-  },
-
-  cleanUp: function (removeFromParent) {
-    Tile.prototype.cleanUp.call(this, removeFromParent);
-    if (this._vertexList$1 != null) {
-      this._vertexList$1 = null;
-    }
-    if (this._childTriangleList$1 != null) {
-      this._childTriangleList$1 = null;
-    }
-    this._subDivided$1 = false;
-    this.demArray = null;
-  }
-};
-
-registerType("PlotTile", [PlotTile, PlotTile$, Tile]);
 
 // wwtlib.TangentTile
